added members, invites and fixed bugs

+++ src/main/java/com/vce/vce/hubs/HubService.java
@@ -1,14 +1,16 @@
 package com.vce.vce.hubs;
 
 import com.vce.vce._shared.model.dto.PageableDTO;
-import com.vce.vce._shared.security.UserContextHolder;
 import com.vce.vce.hubs.dto.CreateOrUpdateHubDTO;
 import com.vce.vce.hubs.dto.HubDTO;
+import com.vce.vce.user.User;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 
 @Service
 @RequiredArgsConstructor
@@ -20,21 +22,6 @@ public class HubService {
     public Page<HubDTO> getAllHubs(PageableDTO pageableDTO) {
         return hubRepository.findAll(pageableDTO.toPageable())
                 .map(hubMapper::toDTO);
-    }
-
-    @Transactional
-    public Hub createHub(CreateOrUpdateHubDTO createHubDTO) {
-        Hub hub = Hub.builder()
-                .name(createHubDTO.name())
-                .owner(UserContextHolder.getCurrentUser())
-                .build();
-
-        return hubRepository.save(hub);
-    }
-
-    @Transactional
-    public HubDTO create(CreateOrUpdateHubDTO createHubDTO) {
-        return hubMapper.toDTO(createHub(createHubDTO));
     }
 
     @Transactional
@@ -55,8 +42,20 @@ public class HubService {
                 .orElseThrow(() -> new EntityNotFoundException("Hub not found"));
     }
 
+    @Transactional(readOnly = true)
+    public Hub findHubById(Long id) {
+        return hubRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Hub not found"));
+    }
+
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, User currentUser) {
+        Hub hub = findHubById(id);
+
+        if (!hub.getOwner().getId().equals(currentUser.getId())) {
+            throw new AccessDeniedException("Insufficient rights");
+        }
+
         hubRepository.deleteById(id);
     }
 }
